requirements.txt
python-dotenv
aiogram
bs4
requests
Flask